+++ internal/global/instruments_test.go
@@ -22,6 +22,7 @@ import (
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/metric/embedded"
 	"go.opentelemetry.io/otel/metric/instrument"
+	"go.opentelemetry.io/otel/metric/noop"
 )
 
 func testFloat64Race(interact func(context.Context, float64, ...attribute.KeyValue), setDelegate func(metric.Meter)) {
@@ -37,7 +38,7 @@ func testFloat64Race(interact func(context.Context, float64, ...attribute.KeyVal
 		}
 	}()
 
-	setDelegate(metric.NewNoopMeterProvider().Meter(""))
+	setDelegate(noop.NewMeterProvider().Meter(""))
 	close(finish)
 }
 
@@ -54,7 +55,7 @@ func testInt64Race(interact func(context.Context, int64, ...attribute.KeyValue),
 		}
 	}()
 
-	setDelegate(metric.NewNoopMeterProvider().Meter(""))
+	setDelegate(noop.NewMeterProvider().Meter(""))
 	close(finish)
 }